requirements.txt
pandas
numpy
streamlit==1.47.1
nltk==3.9.1
scikit-learn==1.7.1
faiss-cpu
sentence-transformers
torch
requests
